# engine/internal/api/tarif.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tarif.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Tarifs"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/tarifs": {
      "get": {
        "operationId": "Tarifs_TarifListing",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TarifListingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Tarifs"
        ]
      },
      "post": {
        "operationId": "Tarifs_TarifCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TarifCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TarifCreateRequest"
            }
          }
        ],
        "tags": [
          "Tarifs"
        ]
      }
    },
    "/api/v1/tarifs/{id}": {
      "get": {
        "operationId": "Tarifs_TarifRead",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TarifReadResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Tarifs"
        ]
      },
      "delete": {
        "operationId": "Tarifs_TarifDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TarifDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Tarifs"
        ]
      }
    },
    "/api/v1/tarifs/{object.id}": {
      "put": {
        "operationId": "Tarifs_TarifUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TarifUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "object.id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TarifsTarifUpdateBody"
            }
          }
        ],
        "tags": [
          "Tarifs"
        ]
      }
    }
  },
  "definitions": {
    "TarifsTarifUpdateBody": {
      "type": "object",
      "properties": {
        "object": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "startDate": {
              "$ref": "#/definitions/v1Date"
            },
            "endDate": {
              "$ref": "#/definitions/v1Date"
            }
          }
        }
      },
      "title": "Updating"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CRUDResult": {
      "type": "object",
      "properties": {
        "rowsAffected": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1Date": {
      "type": "object",
      "properties": {
        "yyyyMmDd": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1TarifCreateRequest": {
      "type": "object",
      "properties": {
        "object": {
          "$ref": "#/definitions/v1TarifInfo"
        }
      },
      "title": "Creating"
    },
    "v1TarifCreateResponse": {
      "type": "object",
      "properties": {
        "crudResult": {
          "$ref": "#/definitions/v1CRUDResult"
        },
        "object": {
          "$ref": "#/definitions/v1TarifInfo"
        }
      }
    },
    "v1TarifDeleteResponse": {
      "type": "object",
      "properties": {
        "crudResult": {
          "$ref": "#/definitions/v1CRUDResult"
        }
      }
    },
    "v1TarifInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "startDate": {
          "$ref": "#/definitions/v1Date"
        },
        "endDate": {
          "$ref": "#/definitions/v1Date"
        }
      }
    },
    "v1TarifListingResponse": {
      "type": "object",
      "properties": {
        "crudResult": {
          "$ref": "#/definitions/v1CRUDResult"
        },
        "objects": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1TarifInfo"
          }
        }
      }
    },
    "v1TarifReadResponse": {
      "type": "object",
      "properties": {
        "crudResult": {
          "$ref": "#/definitions/v1CRUDResult"
        },
        "object": {
          "$ref": "#/definitions/v1TarifInfo"
        }
      }
    },
    "v1TarifUpdateResponse": {
      "type": "object",
      "properties": {
        "crudResult": {
          "$ref": "#/definitions/v1CRUDResult"
        },
        "object": {
          "$ref": "#/definitions/v1TarifInfo"
        }
      }
    }
  }
}